requirements.txt
absl-py==2.3.1
annotated-types==0.4.0
anyio==4.12.0
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.2
coloredlogs==15.0.1
comm==0.2.3
contextlib2==21.6.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
disjoint_set==0.8.0
dm-tree==0.1.9
executing==2.2.1
fastjsonschema==2.21.2
flatbuffers==24.3.25
fonttools==4.61.0
fqdn==1.5.1
future==1.0.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hailo-dataflow-compiler @ file:///home/walter/Downloads/hailo_dataflow_compiler-3.33.0-py3-none-linux_x86_64.whl#sha256=788fddd66d1efbdcb8749b37b06f082f0eefc17dac78dab467a3c96096953155
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.5.0
kiwisolver==1.4.9
lark==1.3.1
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
notebook==7.5.0
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.16.0
onnx-tf==1.6.0
onnxruntime==1.18.0
onnxsim==0.4.36
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pbr==7.0.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prettytable==3.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pwlf==2.5.2
py==1.11.0
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.0.2
pydantic_core==2.1.2
Pygments==2.19.2
pygraphviz==1.14
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
safetensors==0.7.0
scipy==1.12.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-probability==0.20.1
termcolor==3.2.0
terminado==0.18.1
testresources==2.0.2
tflite==2.10.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
verboselogs==1.7
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==2.0.1
xxhash==3.5.0
zipp==3.23.0
